requirements.txt
speechrecognition
pyttsx3
requests
AppOpener
playsound3
google-genai
python-dotenv
